refactor: 베팅 관련 로직은 betting 패키지로 분리
