+++ src/main/java/blackjack/domain/betting/Betting.java
@@ -1,4 +1,4 @@
-package blackjack.domain.game;
+package blackjack.domain.betting;
 
 import java.util.Objects;
