+++ src/main/java/blackjack/domain/betting/BettingTable.java
@@ -1,4 +1,4 @@
-package blackjack.domain.game;
+package blackjack.domain.betting;
 
 import blackjack.domain.participant.Participant;
