+++ src/test/java/blackjack/domain/game/ResultStateTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain.game;
 
+import blackjack.domain.betting.Betting;
 import blackjack.domain.participant.Name;
 import blackjack.domain.participant.Player;
 import blackjack.dto.ResultDto;
